Bug 896664 Split
Fixed incorrect chevron ( > [arrow], issues with the mobile responsive layout of this website. I modified the CSS tags in the file ui.less, where the focus was mainly on the media tags. Specifically, I added 10 to 20 pixels of space between each panel. I also increased the bottom spacing to make the content less cluttered for a mobile user.